Redisassembly with ArrayBoilerplateDescriptions hackily dumped
Redoing disassembly does not preserve addresses, to writing this under
a different name `_v2`, so that the decompilation work is still useful
separately.

I noticed that ArrayBoilerplaceDescriptions in constant pools were
not getting outputted. This is a hacky attempt to do so, but looks
reasonable on inspection. In particular, I wanted to make sure that
some of the decompiled array constants matched what was on the local
binary.