Add SAVE_JSONL_PATH option to record game ticks
This will make it a lot easier to iterate on replicating server logic by
replaying existing games and making sure we produce the same bot logic,
instead of having to relaunch games and hope we hit mismatches with our
bot (which has to search and makes things slow). A follow-up commit will
add a replayer binary.