Optimization: get_aggressive_path -96%
Remove the per-loop sorting of all unseen tiles in pathfinding. This was
a heavy pathfinding bottleneck based on profiling.

This is tricky to get right because the JS implementation has very
unique caveats with the order that it will visit equal-cost tiles,
added documentation to that effect.

While implementing this I was enforcing via asserts that the sorting
approach and optimized approach would iterate in the same tile order,
but this should really be enforced via unit test (to enable future
optimizations), e.g. by refactoring the logic to be a struct where it's
easy to run slow & fast versions step-by-step & compare results.