Optimization: early exit get_aggressive_path -93%
Added a benchmark task where the player is near.
On the task where the player is near, big speedup. No change for task
where player is far (previous benchmark).
This should make a good difference in real play, since the bot is
chasing us it shouldn't always on the other side of the map.