WIP: Implementation of MCTS lib, random search for now
Following https://arxiv.org/pdf/1208.4692 to implement a generalizable
MCTS (and in theory an MCTS grammar that can be searched over to find
search structures that work well!)
Only have implemented 'simulate' for now, but this interestingly allows
us to have iterative sampling with just 'sim(rand_policy)'! This got
a score of ~5k in my test run. Not swapping the search yet, want to make
sure that it consistently gives better results before doing that.