Precompute pathfinding moves (36% search speedup)
This slows down pathfinding grid creation (~5ms), but makes simulation
faster (& thus search faster). Having our first tick be slow is fine,
since the server has leeway when we are slow IIRC, but if needed we can
speed it up as well (e.g. if aggressive precomputation is the
bottleneck, we can compute the 'first move' as part of the
PathfinderState dynamic programming expansion (i.e. the first move of
our "came_from" parent is the same for us).