Add option to run a game without allowing "jumping over" enemies
This makes the search harder, but we still get 10k on each map with it.
Notably, folks who have not reverse engineered the server's code
presumably aren't jumping over enemies, so this makes the bot a bit more
fair in that sense.
Not changing the default `State` creation to use this option (i.e. it
still matches the server logic as-is), but will upload a version where
the python-based bot does no-jump-overs, to see what that looks like.